+++ client-typescript/src/Components/Login.tsx
@@ -1,39 +1,43 @@
-import React, { useState } from 'react';
+import { useState } from 'react';
 import {Link, useNavigate} from 'react-router-dom';
-import {useSetRecoilState} from "recoil";
-import {authState} from "../store/authState.js";
 
-const Signup = () => {
-    const [username, setUsername] = useState('');
-    const [password, setPassword] = useState('');
+interface LoginResponse {
+    _id: string;
+    token: string;
+}
 
-    const handleSignup = async () => {
-        const response = await fetch('http://localhost:3000/auth/signup', {
+const Login = () => {
+    const [username, setUsername] = useState<string>('');
+    const [password, setPassword] = useState<string>('');
+    const navigate = useNavigate();
+
+    const handleLogin = async () => {
+        const response = await fetch('http://localhost:3000/auth/login', {
             method: 'POST',
             headers: { 'Content-Type': 'application/json' },
             body: JSON.stringify({ username, password })
         });
         // Todo: Create a type for the response that you get back from the server
-        const data = await response.json();
+        const data : LoginResponse = await response.json();
         if (data.token) {
             localStorage.setItem("token", data.token)
-            window.location = "/todos";
+            navigate("/todos")
         } else {
-            alert("Error while signing up");
+            alert("invalid credentials");
         }
     };
 
     return (
         <div style={{justifyContent: "center", display: "flex", width: "100%"}}>
             <div>
-                <h2>Signup</h2>
+                <h2>Login</h2>
                 <input type='text' value={username} onChange={(e) => setUsername(e.target.value)} placeholder='Username' />
                 <input type='password' value={password} onChange={(e) => setPassword(e.target.value)} placeholder='Password' />
-                Already signed up? <Link to="/login">Login</Link>
-                <button onClick={handleSignup}>Signup</button>
+                New here? <Link to="/signup">Signup</Link>
+                <button onClick={handleLogin}>Login</button>
             </div>
         </div>
     );
 };
 
-export default Signup;
+export default Login;
