+++ client-typescript/src/Components/Signup.tsx
@@ -1,37 +1,43 @@
-import React, { useState } from 'react';
-import {Link} from 'react-router-dom';
+import { useState } from 'react';
+import {Link, useNavigate} from 'react-router-dom';
 
-const Login = () => {
-    const [username, setUsername] = useState('');
-    const [password, setPassword] = useState('');
+interface SignupResponse {
+    _id: string;
+    token: string;
+}
 
-    const handleLogin = async () => {
-        const response = await fetch('http://localhost:3000/auth/login', {
+const Signup = () => {
+    const [username, setUsername] = useState<string>('');
+    const [password, setPassword] = useState<string>('');
+    const navigate = useNavigate();
+
+    const handleSignup = async () => {
+        const response = await fetch('http://localhost:3000/auth/signup', {
             method: 'POST',
             headers: { 'Content-Type': 'application/json' },
             body: JSON.stringify({ username, password })
         });
         // Todo: Create a type for the response that you get back from the server
-        const data = await response.json();
+        const data : SignupResponse = await response.json();
         if (data.token) {
             localStorage.setItem("token", data.token)
-            window.location = "/todos";
+            navigate("/todos")
         } else {
-            alert("invalid credentials");
+            alert("Error while signing up");
         }
     };
 
     return (
         <div style={{justifyContent: "center", display: "flex", width: "100%"}}>
             <div>
-                <h2>Login</h2>
+                <h2>Signup</h2>
                 <input type='text' value={username} onChange={(e) => setUsername(e.target.value)} placeholder='Username' />
                 <input type='password' value={password} onChange={(e) => setPassword(e.target.value)} placeholder='Password' />
-                New here? <Link to="/signup">Signup</Link>
-                <button onClick={handleLogin}>Login</button>
+                Already signed up? <Link to="/login">Login</Link>
+                <button onClick={handleSignup}>Signup</button>
             </div>
         </div>
     );
 };
 
-export default Login;
+export default Signup;
