+++ client-typescript/src/Components/TodoList.tsx
@@ -1,12 +1,21 @@
-import React, { useContext, useState, useEffect } from 'react';
-import { authState } from '../store/authState.js';
+import { useState, useEffect } from 'react';
+import { useNavigate } from 'react-router-dom';
+import { authState } from '../store/authState';
 import {useRecoilValue} from "recoil";
 
+interface TodoResponse {
+    _id: string;
+    title: string;
+    description: string;
+    done: boolean;
+}
+
 const TodoList = () => {
-    const [todos, setTodos] = useState([]);
+    const [todos, setTodos] = useState<TodoResponse[]>([]);
     const [title, setTitle] = useState('');
     const [description, setDescription] = useState('');
     const authStateValue = useRecoilValue(authState);
+    const navigate = useNavigate();
 
     useEffect(() => {
         const getTodos = async () => {
@@ -46,7 +55,7 @@ const TodoList = () => {
                 <div style={{marginTop: 25, marginLeft: 20}}>
                     <button onClick={() => {
                         localStorage.removeItem("token");
-                        window.location = "/login";
+                        navigate("/login");
                     }}>Logout</button>
                 </div>
             </div>
